woo commerce integration and variable event creation.

+++ inc/admin.class.php
@@ -16,16 +16,17 @@ class mindeventsAdmin {
     $this->default_start_time = (isset($this->options['mindevents_start_time']) ? $this->options['mindevents_start_time'] : '2:00 PM');
     $this->default_end_time = (isset($this->options['mindevents_end_time']) ? $this->options['mindevents_end_time'] : '6:00 PM');
     $this->default_event_color = (isset($this->options['mindevents_event_color']) ? $this->options['mindevents_event_color'] : '#43A0D9');
-    $this->default_event_cost = (isset($this->options['mindevents_event_cost']) ? $this->options['mindevents_event_cost'] : '$25');
+    $this->default_event_cost = (isset($this->options['mindevents_event_cost']) ? $this->options['mindevents_event_cost'] : '25');
 
     add_action( 'add_meta_boxes', array($this, 'add_events_metaboxes' ));
 
-    add_action( 'save_post', array($this, 'save_meta_info'), 10, 2 );
+    add_action( 'save_post_events', array($this, 'save_meta_info'), 10, 2 );
 
 
 
 	}
   static function add_events_metaboxes() {
+    $options = get_option( 'mindevents_support_settings' );
     // add_meta_box( $id, $title, $callback, $page, $context, $priority, $callback_args );
   	add_meta_box(
   		'mindevents_calendar',
@@ -45,6 +46,19 @@ class mindeventsAdmin {
   		'default'
   	);
 
+
+    // if($options['mindevents_enable_woocommerce'] == true && class_exists('woocommerce')) :
+    //   add_meta_box(
+    // 		'mindevents_woo_options',
+    // 		'WooCommerce Options',
+    // 		array('mindeventsAdmin', 'display_woo_options_metabox' ),
+    // 		'events',
+    // 		'side',
+    // 		'default'
+    // 	);
+    // endif;
+
+
   }
 
 
@@ -58,7 +72,7 @@ class mindeventsAdmin {
     if ( !isset( $_POST['mindevents_event_meta_nonce'] ) || !wp_verify_nonce( $_POST['mindevents_event_meta_nonce'], basename( __FILE__ ) ) )
       return $post_id;
 
-
+    update_post_meta( $post_id, 'event_defaults', $_POST['event']);
 
     $field_key = 'event_meta';
     /* Get the posted data and sanitize it for use as an HTML class. */
@@ -69,10 +83,9 @@ class mindeventsAdmin {
       endforeach;
     endif;
 
+
     return $post_id;
   }
-
-
 
 
   static function display_event_options_metabox() {
@@ -103,9 +116,6 @@ class mindeventsAdmin {
   }
 
 
-
-
-
   static function display_calendar_metabox($post) {
     echo '<div class="mindevents_meta_box mindevents-forms" id="mindevents_meta_box">';
       echo '<h3>Occurance Options</h3>';
@@ -127,12 +137,28 @@ class mindeventsAdmin {
     echo '</div>';
   }
 
+  static function display_woo_options_metabox() {
+    wp_nonce_field( basename( __FILE__ ), 'mindevents_event_meta_nonce' );
+    echo '<div class="mindevents_meta_box mindevents-forms" id="mindevents_meta_box">';
+      
+
+      echo '<div class="form-section">';
+        echo '<p class="label">';
+          echo '<input type="checkbox" name="woo_create_variations" id="woo_create_variations" value="1" ' . checked(get_post_meta(get_the_ID(), 'woo_create_variations', true), '1', false) . '>';
+          echo '<label for="woo_create_variations">Create Product Variation for Each event Occurance?</label>';
+        echo '</p>';
+      echo '</div>';
+
+
+    echo '</div>';
+
+  }
 
 
 
   private function get_time_form() {
-    $defaults = get_post_meta(get_the_ID(), 'defaults', true);
-
+    $defaults = get_post_meta(get_the_ID(), 'event_defaults', true);
+    $options = get_option( 'mindevents_support_settings' );
     echo '<fieldset id="defaultEventMeta" class="event-times mindevents-forms">';
       echo '<div class="time-block">';
         echo '<div class="form-section">';
@@ -141,7 +167,7 @@ class mindeventsAdmin {
         echo '</div>';
         echo '<div class="form-section">';
           echo '<p class="label"><label for="endtime">Event Occurence End</label></p>';
-          echo '<input type="text" class="timepicker" name="event[endtime]" id="endtime" value="' . (isset($defaults['endtime']) ? $defaults['endtime'] : $this->default_start_time) . '" placeholder="">';
+          echo '<input type="text" class="timepicker" name="event[endtime]" id="endtime" value="' . (isset($defaults['endtime']) ? $defaults['endtime'] : $this->default_end_time) . '" placeholder="">';
         echo '</div>';
 
         echo '<div class="form-section">';
@@ -153,35 +179,63 @@ class mindeventsAdmin {
           echo '<p class="label"><label for="eventDescription">Short Description</label></p>';
           echo '<textarea type="text" name="event[eventDescription]" id="eventDescription" value="' . (isset($defaults['eventDescription']) ? $defaults['eventDescription'] : '') . '" placeholder="">' . (isset($defaults['eventDescription']) ? $defaults['eventDescription'] : '') . '</textarea>';
         echo '</div>';
-
-
-        echo '<h3 class="offers-title">Tickets</h3>';
-        echo '<div class="offer-options" id="allOffers">';
-
-          echo '<div class="single-offer">';
-            echo '<div class="form-section">';
-              echo '<p class="label"><label for="eventLinkLabel">Ticket Label</label></p>';
-              echo '<input type="text" name="event[offerlabel][]" id="eventLinkLabel" value="' . (isset($defaults['eventLinkLabel']) ? $defaults['eventLinkLabel'] : 'General Admission') . '" placeholder="">';
-            echo '</div>';
+        
+        if($options['mindevents_enable_woocommerce'] == true && class_exists('woocommerce')) :
+            //add hiden inpuit
+            echo '<input type="hidden" name="event[wooLinked]" id="wooLink" value="1">';
 
             echo '<div class="form-section">';
-              echo '<p class="label"><label for="eventCost">Price</label></p>';
-              echo '<input type="text" name="event[offerprice][]" id="eventCost" value="' . (isset($defaults['eventCost']) ? $defaults['eventCost'] : '') . '" placeholder="">';
+              echo '<p class="label"><label for="eventProductID">Ticket Button Text</label></p>';
+              echo '<input type="text" name="event[wooLabel]" id="eventLinkedProduct" value="' . (isset($defaults['wooLabel']) ? $defaults['wooLabel'] : '') . '" placeholder="">';
             echo '</div>';
 
             echo '<div class="form-section">';
-              echo '<p class="label"><label for="eventLink">Link</label></p>';
-              echo '<input type="text" name="event[offerlink][]" id="eventLink" value="' . (isset($defaults['eventLink']) ? $defaults['eventLink'] : '') . '" placeholder="">';
+              echo '<p class="label"><label for="eventProductID">Linked Product</label></p>';
+              echo '<input type="number" name="event[wooLinkedProduct]" id="eventLinkedProduct" value="' . (isset($defaults['wooLinkedProduct']) ? $defaults['wooLinkedProduct'] : '') . '" placeholder="">';
+              echo '<p class="description">If left blank a new product will be created.</p>';
+            echo '</div>';
+            
+            echo '<div class="form-section">';
+              echo '<p class="label"><label for="eventCost">Event Cost</label></p>';
+              echo '<input type="number" name="event[wooPrice]" id="eventCost" value="' . (isset($defaults['wooPrice']) ? $defaults['wooPrice'] : '') . '" placeholder="100">';
+              echo '<p class="description">This will be the default cost for all tickets, it can be changed on the WooCommerce product.</p>';
             echo '</div>';
 
-            echo '<div class="add-offer">';
-              echo '<span>+</span>';
+            echo '<div class="form-section">';
+              echo '<p class="label"><label for="wooStock">Event Stock</label></p>';
+              echo '<input type="number" name="event[wooStock]" id="wooStock" value="' . (isset($defaults['wooStock']) ? $defaults['wooStock'] : '4') . '" placeholder="4">';
+              echo '<p class="description">This will be the stock limit for the associated product variation, it will be ignored for simple products.</p>';
             echo '</div>';
+            
+          
+        else : 
+          echo '<h3 class="offers-title">Tickets</h3>';
+          echo '<div class="offer-options" id="allOffers">';
+            echo '<div class="single-offer">';
+              echo '<div class="form-section">';
+                echo '<p class="label"><label for="eventLinkLabel">Ticket Label</label></p>';
+                echo '<input type="text" name="event[offerlabel][]" id="eventLinkLabel" value="' . (isset($defaults['eventLinkLabel']) ? $defaults['eventLinkLabel'] : 'General Admission') . '" placeholder="">';
+              echo '</div>';
+
+              echo '<div class="form-section">';
+                echo '<p class="label"><label for="eventCost">Price</label></p>';
+                echo '<input type="text" name="event[offerprice][]" id="eventCost" value="' . (isset($defaults['eventCost']) ? $defaults['eventCost'] : '') . '" placeholder="">';
+              echo '</div>';
+
+              echo '<div class="form-section">';
+                echo '<p class="label"><label for="eventLink">Link</label></p>';
+                echo '<input type="text" name="event[offerlink][]" id="eventLink" value="' . (isset($defaults['eventLink']) ? $defaults['eventLink'] : '') . '" placeholder="">';
+              echo '</div>';
+
+              echo '<div class="add-offer">';
+                echo '<span>+</span>';
+              echo '</div>';
+            echo '</div>';
+
+
+
           echo '</div>';
-
-
-
-        echo '</div>';
+        endif;
 
 
       echo '</div>';
